This patch adds various optimizations extracted from the thread_local_allocator work
- make AGAS properly (fast-)resolve local only ids
- avoid reading atomic in future state  more than necessary
- marking certain hot-spot functions as force-inline
- extend future_access_customization_point to support addref control while creating a future from a shared state
- unwrapping_result_policy now holds id by reference
- flyby: fix warnings in when_each test
- flyby: split partitioned_vector_exclusive_scan test

This also adds overloads for APIs that produce futures taking an allocator (dataflow, make_ready_future, etc.)

+++ tests/unit/parallel/segmented_algorithms/partitioned_vector_exclusive_scan2.cpp
@@ -350,7 +350,7 @@ void exclusive_scan_tests(std::vector<hpx::id_type>& localities)
 int main()
 {
     std::vector<hpx::id_type> localities = hpx::find_all_localities();
-    exclusive_scan_tests<long long>(localities);
+    exclusive_scan_tests<double>(localities);
 
     return hpx::util::report_errors();
 }
